+++ cropduster/standalone/views.py
@@ -30,7 +30,7 @@ class CropDusterStandaloneIndex(CropDusterIndex):
             (preview_w, preview_h) = self.preview_size
             standalone = StandaloneImage.objects.get_from_file(self.image_file.name,
                 upload_to=self.upload_to, preview_w=preview_w, preview_h=preview_h)
-        db_image = standalone.image.cropduster_image
+        db_image = standalone.image.related_object
         if not getattr(db_image, 'pk', None):
             raise Exception("Image does not exist in database")
         if not db_image.image and standalone.image.name:
