+++ cropduster/templates/cropduster/custom_field.html
@@ -4,9 +4,9 @@
 <input type="text"
        id="{{ final_attrs.id }}"
        name="{{ final_attrs.name }}"
-       value="{{ image_value }}"
+       value="{{ file_value }}"
        class="cropduster-data-field cropduster-text-field"
-       data-sizes="{{ final_attrs.sizes }}"
+       data-sizes="{{ sizes }}"
        data-upload-to="{{ upload_to }}"
        data-prefix="{{ prefix }}"/>
